An inbox form endpoint.
In creating the inbox controller, there was a need for a different method of testing.
Rather than testing the request/response lifecycle of express, I needed to
test the hypermedia specifically. The problem was that I was having trouble getting
the protocal and the hostname off the response. Extracting the creation of the hypermedia
out of the route definition and into a function on it's own enabled passing parameters
into the construction of a hypermedia response, and more testable.
Perhaps there still needs to be a test that checks that inboxController.constructHypermedia is called when you
hit the route? Perhaps this approach is misguided? Looking for thoughts.

S-47060