WIP: Begin using eventStoreClient helper in queryController. Test whether eventStoreClient.streams.get is called appropriately when asking for all workitems.
The use of the helper enables us to no longer new up EventStore in the
controller and makes it mockable/stubbable/spyable.

This is a WIP, because we are in the process of converting the tests over
to use the newly mockable eventstore client via proxyquire. Former/older
tests have temporarily been commented out, and will be brought back one
at a time.

S-50752