No longer make this particular test care about the timeFormatted field in the response.
Need to remove the 'timeFormatted' field from the response body as
we can not rely on that value staying constant as time moves on. It will
change because we are using the timeago package to transform it into values
like '3 months ago'. But once a month increments, then our assertion will be wrong
because the actual value at runtime will then be '4 months ago'

S-50918