Added back .deep :-D
Note: for some tests using just .equals will work, but ONLY if the data inside the object doesn't exactly match. When the data inside the objects does exactly match, it appears that .equals then fails it because they do not point to the same object in memory. I guess.

S-51903