pull the smoke tests back in line.
Needed to add a delay in sending a response after creating a digest so that
creating an inbox would be able to validate that the digest exists.

Eventstore seems to take a second or so to get the projection available for
the digest. We took the simplest answer for now for that issue, and will revisit
it if necessary.

S-50918