Remove the link to the inboxes resource for creating an inbox, as well as the digestId property.
Neither of these are being used in any workflow, and therefore may be
unnecessary. Removing them for now until they earn their way back in.

As for the digest Id, there is a link to the digest parent in order to get
more information about that digest.

As for the link to inboxes... if it were to stay then in my opinion it would
be to use the digestId to create a new inbox associated to the same
digest for the inbox that you just retrieved information on. We're not
in need of this yet I don't think.

S-51313