S-52445 - pull in changes to make CommitStream work with EventStore 3.0.2.
Using the tryharder parameter when retreiving streams. From eventstore's
documentation it will:

"work even harder to try to parse and reformat json from an event to
allow it to be returned in the feed"

The scructure of the payload that comes back from eventstore changed and
lives right in .data rather than content.data.