S-51782: use an angular directive (??whoknew) to handle resizing the overlay when the window resizes.
this was inspired by this article:
http://microblog.anthonyestebe.com/2013-11-30/window-resize-event-with-angular/

i hacked out pieces little by little that didn't seem to be needed.

included in this commit is a jQuery based approach but it is commented out,
just wanted to keep it in for the time being to share with folks on Monday.

also included a .overlay class on the image itself and modified css and
selectors to use that instead of the img tag.