remove no longer necessary test and ad some words of caution when modifying a couple of variables in this file.
These variables are representative of the files that get read in when
sending files down to the client in the case of SVN. They are
tempermental because there are lots of apostrophe's (') and double quotes(")
and curly braces ({}) that need to get escaped to have the test succesfully run
when comparing the strings. lovely fun.