D-11614 D-11613 Change the approach from validating the url to cleaning it
We've gone through many gyrations on how this should be handled. We have
arrived here after trying several things like:

1. utilizing regexes to ensure we have a valid url (this is difficult and
prone to missing cases)
2. Html Encoding it (does not work for an anchor tag in a successful case)
3. Cleaning the url of known things that should not be there for our
current failure cases.

Option 3 is what we have gone with for now.

We have also introduced the first tested code for the client ended
javascript, which feels like a big win for us. It's been a hole in
how we have practiced development on the client here, and it feels good
to have a path forward... even if that path is one that needs improvements.