S-60862: A HACK on a HACK. Switch to use a different icon for a VsoTfvc inbox.
The hack that already exists is the icon-vsogit-selected-32x32 icon is
not the icon for git, it is the icon for VSTS.

This hack further compounds the issue as we now include the VsoTfvc inbox
to use the same thing.

Will need to come back around and fix the naming of the icon and set the
if to include both of them VsoGit and VsoTfvc to be associated to the new icon
prefereable named icon-vsts-selected-32x32.

Not sure if the current icon name is being used anywhere else, and I have to
run for the day.