S-60862: Set up a section for VsoTfvc and move everything else to a section that is not VsoTfvc.
We know, this is an ugly thing. We're trying to release and need to make
forward progress quickly... this is not how we would like to do it, as the
template for sure get's more complex with this commit.

We currently use a version of express-handlebars that pulls in a version of
handlebars that does not support if-elseif-else functionality. We were
relying on that type of basic functionality being there, but it unfortunately
is not in our case. Pulling that in the week of the release and regression
testing that piece is not a risk I think that we are willing to take
at this moment, but have posed the question.

This commit is an alternative to that, and is supported by our current
version.