S-62662 Removed util.inspect(err, ...) line for performance gain
This call was taking exorbitant amounts of time for certain exception err
objects to be processed. It does not buy us any additional information
besides a highly verbose and often redundant (including circular links!)
information.

Locally, it has a gigantic effect on the throughput, documented in this
wiki page: https://github.com/openAgile/CommitStream.Web/wiki/CS-Performance-Optimization-research.

Testing on v1-cs-test It takes about 55 seconds for 50 concurrent curl.exe
processes to execute 10 requests each. Locally it's about 9 seconds. It's
hard to say how long it takes in Azure with the old code in place, but
trying to do 50 curls for just 1 single request each took 3:54! I won't
assume that we would just multiple that by 10 know how long it would take
for 10 each, but clearly it's an issue. Running the same lower dose
without the line averages between 3 and 10 seconds.