S-62662 Added catchAsyncErrors module to standardize .catch to forward errors
A big problem in the code was that in several key areas we were making calls
to eventStoreClient without utilizing the returned Promise object. Instead,
we were relying on the code in the .then() handler to format responses or
attach items to the contextual req object.

Additionally, we were sometimes losing thrown exceptions because the
express-domain-middleware module that we were relying on to trap uncaught
exceptions is not reliable (and is now deprecated as the domain feature is
going to be removed from a future release of Node).

Because of that, we needed a way to get back those exceptions and avoid
relying on the domain as a last chance way to grab them. With the help
of some articles, this module can be used to wrap around any middleware
that returns a Promise. It then attached a standardized catch handler that
calls the supplied next function, which represents that csErrorHandler
itself in our app.

When testing this we have yet to see a case where ANY thrown exceptions
fail to be processed and handled by the csErrorHandler. This includes
running up to 20000 errors in a short time period. Errors can be reproduced
by running the appropriate scripts in https://github.com/openAgile/CommitStream.Web.Tests/tree/master/src/tests/performance.

References:
https://nodejs.org/api/domain.html
https://expressjs.com/en/advanced/best-practice-performance.html#use-promises
http://madole.xyz/error-handling-in-express-with-async-await-routes/
https://strongloop.com/strongblog/async-error-handling-expressjs-es7-promises-generators/?_ga=2.104300101.459885209.1495130589-1831304197.1493136526