S-62662 Applied catchAsyncErrors wrapper on validator middleware
In order to catch thrown exceptions directly, instead of relying on
express-domain-middleware, the wrapper applies the .catch and calls next,
as described in the previous commit.

This only required making sure each of the validator middleware functions
actually returned the Promise from eventStoreClient, instead of just calling
it and relying on the .then handler.

The differences in performance here is dramatic.

Locally, when running without the wrapper here are three run results:


time ./runner.sh 100 "./query-non-existent-instance.sh 5"

1: 6m4.254s
2: 8m4.593s
3: 6m4.366s

And, running the same command after applying the wrapper:

1: 0m7.509s
2: 0m7.373s
3: 0m8.284s

That's an average of 404.404 seconds in the unwrapped case. (Yeah,
I'd say performance is NOT FOUND NOT FOUND)

And, an average of 7.722 seconds in the wrapped case, for a whopping
difference of 52.37 times faster!