S-62662 Wrap commitsCreate with catchAsyncErrors
While tests show that, because the MalformedPushEventError is captured
prior to invoking eventStoreClient, the incoming invalid payloads are
not subject to hanging the process, once the code does invokve the
client, it is prone to hang if the Promise code throws an exception. I
simulated this by injecting a failure into the .then() handler.

Results:

Without wrapper:
time ./runner.sh 70 "./post-bitbucket-pull-request.sh 10"  -->  0m14.605s (all 700)

With wrapper, the same command --> 0m15.410s (all 700)

Thus, there is slight overhead in this case, but the benefts far outweigh
the costs when it comes to catching eventStoreClient exceptions or Promise
failures, shown here:

Without wrapper:
time ./runner.sh 10 "./post-github-commits.sh 10" --> 2m2.809s
(only 97 errors caught by domain middleware)

And, for 20 by 10 --> 4m1.363s (only 187 errors caught)

By contrast with the wrapper for GitHub:

20 by 10 --> 0m4.105s  (all 200)
40 by 10 --> 0m7.517s  (all 400)
80 by 10 --> 0m13.465s (all 800)

While there are many additional routes in the system that will need wrapping
in order to maximize the benefit of Promise rejection handling, and minimize
the latency due to the erratic nature of the domain middleware, this commit
concludes the first batch for a "patch" that will handle the vast majority of
current runtime situations on the POST side.