Show sub-family icons
Issue: When listing the global repositories in read-only mode
(for example: Global option in TeamRoom CommitStream settings UI), the
sub-family icons are not shown. This is inconsistent with the screens
where the list can be edited (for example, Custom option in TeamRoom
CommitStream settings UI).

Solution: Show sub-family icons in all screens (where the repository is a sub-family).