WIP - DO NOT RELEASE WITHOUT READING MORE HERE
No clue what this is

You have to love thrash. I came back to the codebase after having been
moved onto other things long long long ago. I'm not sure what any of this
is doing or why I even have this branch, but... in the event that this is
actually usable code for something, I'm pushing it up.